Rename `boot` to `Bootloader`
